[#22] telegrambot: Refactor subpackage
Refactor the 'telegrambot' subpackage into the following files:
* telegrambot.go: Contains the telegram bot logic and functions to start
  the bot.
* callback-query.go: Responsible for handling callback queries.
* message.go: Responsible for handling regular messages.
* searches.go: Processes search-related messages and callback queries.
* utils.go: Contains utility functions.
* constants.go: Contains constants (e.g. cities).

Adds documentation to all functions and types of the 'telegrambot' subpackage.
