+++ telegrambot/callback-query.go
@@ -1,58 +1,35 @@
 package telegrambot
 
 import (
-	"apartment-parser/database"
 	"database/sql"
-	tgbotapi "github.com/go-telegram-bot-api/telegram-bot-api/v5"
 	"log"
-	"strconv"
 	"strings"
+
+	tgbotapi "github.com/go-telegram-bot-api/telegram-bot-api/v5"
 )
 
-func ProcessCallbackQuery(bot *tgbotapi.BotAPI, update tgbotapi.Update, search_db *sql.DB) {
+// Process callback query.
+// Verifies the callback query data and calls the appropriate function.
+//
+// Parameters:
+//
+//	bot: Telegram bot instance.
+//	update: Telegram update instance.
+//	search_db: Search database instance.
+func processCallbackQuery(bot *tgbotapi.BotAPI, update tgbotapi.Update, search_db *sql.DB) {
 	data := strings.Split(update.CallbackQuery.Data, "|")
+	switch data[0] {
 
-	if len(data) == 1 {
-		switch update.CallbackQuery.Data {
-		case "create":
-			ListCities(bot, update, search_db)
+	case "remove_msg":
+		removeUpdateQueryMessage(bot, update)
+		return
 
-		case "cleanup":
-			deleteMessage := tgbotapi.NewDeleteMessage(update.CallbackQuery.Message.Chat.ID, update.CallbackQuery.Message.MessageID)
-			_, err := bot.Send(deleteMessage)
-			if err != nil {
-				log.Println(err)
-			}
-		}
-	} else if len(data) == 2 {
-		switch data[0] {
-		case "search":
-			search_id, err := strconv.Atoi(data[1])
-			if err != nil {
-				log.Println(err)
-			} else {
-				listSearchInfo(bot, update, search_db, search_id)
-			}
+	case "search":
+		processSeachAction(bot, update, search_db)
 
-		case "delete":
-			search_id, err := strconv.Atoi(data[1])
-			if err != nil {
-				log.Println(err)
-			}
-			err = database.DeleteSearch(search_db, int64(search_id))
-			if err != nil {
-				log.Println(err)
-			}
-
-		case "new-search":
-			handleNewSearchCity(bot, update, search_db, data[1])
-		}
+	default:
+		log.Println("Unknown callback query data: ", data[0])
 	}
 
-	// Delete callback query
-	deleteMessage := tgbotapi.NewDeleteMessage(update.CallbackQuery.Message.Chat.ID, update.CallbackQuery.Message.MessageID)
-	_, err := bot.Send(deleteMessage)
-	if err != nil {
-		log.Println(err)
-	}
+	removeUpdateQueryMessage(bot, update)
 }
