+++ telegrambot/constants.go
@@ -1,0 +1,64 @@
+package telegrambot
+
+// List of cities supported by the bot
+var cities = []City{
+	{
+		Name: "Białystok",
+		Code: "bialystok",
+	},
+	{
+		Name: "Bydgoszcz",
+		Code: "bydgoszcz",
+	},
+	{
+		Name: "Gdańsk",
+		Code: "gdansk",
+	},
+	{
+		Name: "Gdynia",
+		Code: "gdynia",
+	},
+	{
+		Name: "Katowice",
+		Code: "katowice",
+	},
+	{
+		Name: "Kielce",
+		Code: "kielce",
+	},
+	{
+		Name: "Kraków",
+		Code: "krakow",
+	},
+	{
+		Name: "Lublin",
+		Code: "lublin",
+	},
+	{
+		Name: "Łódź",
+		Code: "lodz",
+	},
+	{
+		Name: "Poznań",
+		Code: "poznan",
+	},
+	{
+		Name: "Radom",
+		Code: "radom",
+	},
+	{
+		Name: "Rzeszów",
+		Code: "rzeszow",
+	},
+	{
+		Name: "Szczecin",
+		Code: "szczecin",
+	},
+	{
+		Name: "Wrocław",
+		Code: "wroclaw",
+	},
+	{
+		Name: "Warszawa",
+		Code: "warszawa",
+	}}
